Added (a mostly) debugged Reviews Function Suite
I have all the functions added: add_review, del_review, generate_random_review, print_dummy_reviews, print_all_reviews. BUT, I cannot for the life of me understand why the reviews aren't working. I encoded support for the reviews to only be added to existing listing ID's, but no matter what, the program returns and says that the ID doesn't exist and skips it. I checked, and most of the ones it picks are ones that DO exist. I tried print statements galore and casting but it still skips them to no avail. I need a second pair of eyes on this.