Successfully debugged Reviews
Debugged a few casting issues with the add_review function. Finally made it so that reviews get added correctly. A review can only be added if one does not already exist for a listing. If so, the review does not get added. That will be its own function implemented at a later date, edit_review.

Co-Authored-By: Peter Murphy <156000804+murphype@users.noreply.github.com>